+++ composeApp/src/commonMain/kotlin/com/nacchofer31/randomboxd/random_film/presentation/viewmodel/RandomFilmViewModel.kt
@@ -4,11 +4,14 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.nacchofer31.randomboxd.core.domain.DispatcherProvider
 import com.nacchofer31.randomboxd.core.domain.ResultData
+import com.nacchofer31.randomboxd.random_film.domain.model.UserName
 import com.nacchofer31.randomboxd.random_film.domain.repository.RandomFilmRepository
+import com.nacchofer31.randomboxd.random_film.domain.repository.UserNameRepository
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.filterIsInstance
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.flow
@@ -18,15 +21,27 @@ import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
+import kotlinx.coroutines.launch
 
 @Suppress("OPT_IN_USAGE")
 class RandomFilmViewModel(
     private val repository: RandomFilmRepository,
+    private val userNameRepository: UserNameRepository,
     private val dispatchers: DispatcherProvider,
 ) : ViewModel() {
     private val actions = MutableSharedFlow<RandomFilmAction>(replay = 1, onBufferOverflow = BufferOverflow.DROP_OLDEST)
 
     private val internalState = MutableStateFlow(RandomFilmState())
+
+    val userNameList: StateFlow<List<UserName>> =
+        userNameRepository
+            .getAllUserNames()
+            .stateIn(
+                scope = viewModelScope,
+                started = SharingStarted.WhileSubscribed(5000),
+                initialValue = emptyList(),
+            )
+
     internal val state =
         internalState
             .stateIn(
@@ -40,7 +55,9 @@ class RandomFilmViewModel(
             .filterIsInstance<RandomFilmAction.OnSubmitButtonClick>()
             .flatMapLatest {
                 flow {
-                    val result = repository.getRandomMovie(internalState.value.userName)
+                    val userName = internalState.value.userName
+                    userNameRepository.addUserName(userName)
+                    val result = repository.getRandomMovie(userName)
                     emit(result)
                 }.onStart {
                     internalState.update { it.copy(isLoading = true) }
@@ -68,6 +85,7 @@ class RandomFilmViewModel(
             is RandomFilmAction.OnSubmitButtonClick -> actions.tryEmit(action)
             is RandomFilmAction.OnClearButtonClick -> clearSelectedFilm()
             is RandomFilmAction.OnUserNameChanged -> updateUserNameValue(action.username)
+            is RandomFilmAction.OnRemoveUserName -> viewModelScope.launch { userNameRepository.deleteUserName(action.userName) }
             else -> Unit
         }
     }