+++ composeApp/src/iosMain/kotlin/com/nacchofer31/randomboxd/MainViewController.kt
@@ -1,7 +1,9 @@
 package com.nacchofer31.randomboxd
 
+import androidx.compose.runtime.remember
 import androidx.compose.ui.window.ComposeUIViewController
 import com.nacchofer31.randomboxd.app.RandomBoxdApp
+import com.nacchofer31.randomboxd.database.getUserNameDatabase
 import com.nacchofer31.randomboxd.di.initKoin
 
 fun mainViewController() =
@@ -9,4 +11,10 @@ fun mainViewController() =
         configure = {
             initKoin()
         },
-    ) { RandomBoxdApp() }
+    ) {
+        val dao =
+            remember {
+                getUserNameDatabase().userNameDao()
+            }
+        RandomBoxdApp(dao)
+    }